feat: 객체 지향 추가 적용 - dataService 및 launchService 정의
